gpio output wip
